Split projects as modules
