+++ Core/src/main/java/put/io/black/java/core/app/ScenarioApplication.java
@@ -1,0 +1,20 @@
+package put.io.black.java.core.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Main Application code to ensure run SpringApplication
+ * @see SpringApplication
+ */
+@SpringBootApplication(scanBasePackages = {"put.io.black.java.core.rest"})
+public class ScenarioApplication {
+
+    /**
+     * Run SpringApplication with our ScenarioApplication class
+     * @param args Program arguments
+     */
+    public static void main(String[] args) {
+        SpringApplication.run(ScenarioApplication.class, args);
+    }
+}
