+++ Core/src/main/java/put/io/black/java/core/logic/Visitable.java
@@ -1,0 +1,14 @@
+package put.io.black.java.core.logic;
+
+/**
+ * Interface Visitable
+ * @see Visitor
+ */
+public interface Visitable {
+
+    /**
+     * Accept visitor check
+     * @param visitor Visitor - guest
+     */
+    public void accept(Visitor visitor);
+}
