+++ Core/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+logging.level.root= WARN
+logging.level.put.io.black.java.core= DEBUG
