--- src/main/java/put/io/black/java/logic/Visitable.java
@@ -1,14 +1,0 @@
-package put.io.black.java.logic;
-
-/**
- * Interface Visitable
- * @see Visitor
- */
-public interface Visitable {
-
-    /**
-     * Accept visitor check
-     * @param visitor Visitor - guest
-     */
-    public void accept(Visitor visitor);
-}
