--- src/main/java/put/io/black/java/logic/Visitor.java
@@ -1,14 +1,0 @@
-package put.io.black.java.logic;
-
-/**
- * Interface Visitor
- * @see Visitable
- */
-public interface Visitor {
-
-    /**
-     * Visit (check) node
-     * @param node Node to check
-     */
-    public void visit(Node node);
-}
