--- src/main/resources/application.properties
@@ -1,2 +1,0 @@
-logging.level.root= WARN
-logging.level.put.io.black.java= DEBUG
